👌 IMPROVE: Seperate course reviews and minor ui imp
