+++ src/components/CourseFinder/CourseSearch.jsx
@@ -74,10 +74,9 @@ const StyledInput = styled(Input)`
   height: 2rem;
   padding: 0 0.75rem;
   border: 0;
-  border-bottom: 2px solid ${({ theme }) => theme.darksecondary};
   border-radius: 0.5rem;
   background: ${({ theme }) => theme.secondary};
-  box-shadow: 0 0 0.7rem rgba(0, 0, 0, 0.2), 0 5px 0.5rem rgba(0, 0, 0, 0.2);
+  box-shadow: 0 8px 5px rgba(0, 0, 0, 0.3);
 
   .ant-input {
     padding: 0 0.25rem;
@@ -91,9 +90,5 @@ const StyledInput = styled(Input)`
     width: 0.875rem;
     height: 0.875rem;
     color: lightgray;
-  }
-
-  &:not(.ant-input-affix-wrapper-disabled):hover {
-    border-color: ${({ theme }) => theme.darksecondary};
   }
 `
